init　norm対応
