+++ src/init/init_mlx.c
@@ -1,0 +1,52 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   init_mlx.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ootsuboyoshiyuki <ootsuboyoshiyuki@stud    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/11/15 15:47:36 by ootsuboyosh       #+#    #+#             */
+/*   Updated: 2024/11/28 14:14:14 by ootsuboyosh      ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "init.h"
+
+void	init_graphic(t_game *game)
+{
+	game->mlx = mlx_init();
+	if (!game->mlx)
+	{
+		perror("Error initializing MinilibX");
+		exit(EXIT_FAILURE);
+	}
+	game->win = mlx_new_window(game->mlx, WINDOW_WIDTH, WINDOW_HEIGHT, "cub3D");
+	if (!game->win)
+	{
+		perror("Error creating window");
+		exit(EXIT_FAILURE);
+	}
+}
+
+void	init_image(t_game *game)
+{
+	game->img.img = mlx_new_image(game->mlx, WINDOW_WIDTH, WINDOW_HEIGHT);
+	if (!game->img.img)
+	{
+		perror("Error creating image");
+		exit(EXIT_FAILURE);
+	}
+	game->img.addr = mlx_get_data_addr(game->img.img, &game->img.bits_per_pixel,
+			&game->img.line_length, &game->img.endian);
+	if (!game->img.addr)
+	{
+		perror("Error getting image address");
+		exit(EXIT_FAILURE);
+	}
+}
+
+void	init_mlx(t_game *game)
+{
+	init_graphic(game);
+	init_image(game);
+}
